delete analytics
